fix(stella-cli): split prompt.rs rather than grandfather it, and unred main
#3035 pushed `crates/stella-cli/src/agent/prompt.rs` to 1503 lines and the
file-size ratchet failed it. The failure was cleared by `make
file-size-update`, which wrote `+1503 crates/stella-cli/src/agent/prompt.rs`
into `scripts/file-size-baseline.txt` — a new grandfathered entry for a file
that postdates the guard. That is the one thing the guard's own failure text
rules out ("Do not add a baseline entry: the baseline only covers files that
predate the guard, and this is the rule that stops the tree getting worse"),
and AGENTS.md § "God files" states the same in the other direction: the
baseline accepts no new entries; a file approaching the limit gets split.

It also left `main` red. `check-god-files.sh` cross-checks every baseline
entry against the god-file table in AGENTS.md and the matching crate README
(#1435), and neither names `src/agent/prompt.rs`. Verified against `main` at
dce1c25af:

    check-god-files: FAIL — AGENTS.md's god-file table omits stella-cli god
    file(s): src/agent/prompt.rs
    check-god-files: FAIL — crates/stella-cli/README.md omits stella-cli god
    file(s): src/agent/prompt.rs

So take the split the guard asked for. The inline `mod tests` moves verbatim
to `agent/prompt/tests.rs`, beside the `parity` submodule that already
establishes the shape: prompt.rs 1503 -> 840, the new file 672, and the
baseline entry is deleted rather than kept. Both guards pass, and no prose
needed updating precisely because the file stopped being a god file.

The split is the right half to move. What makes prompt.rs grow is the prose
above each clause explaining why it is worded as it is — #3035's own overage
was eight lines of doc comment explaining why the escalation clause names no
tool — and that is the half worth the ceiling.

All 17 tests move unchanged and still run as `agent::prompt::tests::*` (31
passing under `agent::prompt::`, counting `parity`). A move between files, so
no test is removed from the tree. Four comments in `parity.rs` and one in
`prompt.rs` justified living there by prompt.rs sitting "within a few lines of
the 1500-line ratchet"; that is false at 840 lines, so each now cites #2237 —
which consolidates prompt-content pins there and stands on its own — instead
of a ceiling that would mislead the next maintainer into thinking prompt.rs
has no room.

Verified: check-file-size OK, check-god-files OK, cargo fmt --check,
clippy -p stella-cli --all-targets -D warnings, and the prompt tests. The
crate's integration binaries were not built locally.

Both prompts verified byte-identical against main at dce1c25af by capturing
each const to a file in both trees and diffing: SYSTEM_PROMPT sha256
6b77b5e279f174f43719ee9d8d2af7d4715e19cfe9fccbd6ca96107b37b0318f and
PIPELINE_SYSTEM_PROMPT sha256
e592ea9b4cba1ffb6c9c3b7a6085a8aaca4d01e10599e1eddfcea76f97055de8 on both
sides, so the cached prefix invariant 7 keys on is untouched.

Closes #2985
